Fix chat lifecycle composition ownership
