+++ scripts/file-length-baseline.json
@@ -16,7 +16,7 @@
   "src/backend/services/session/service/acp/acp-runtime-manager.test.ts": 2727,
   "src/backend/services/session/service/acp/acp-runtime-manager.ts": 1456,
   "src/backend/services/session/service/acp/codex-app-server-adapter/codex-app-server-acp-adapter.test.ts": 3423,
-  "src/backend/services/session/service/chat/chat-message-handlers.service.test.ts": 1219,
+  "src/backend/services/session/service/chat/chat-message-handlers.service.test.ts": 1217,
   "src/backend/services/session/service/chat/chat-message-handlers/handlers/load-session.handler.test.ts": 1576,
   "src/backend/services/session/service/lifecycle/session.config.service.test.ts": 1159,
   "src/backend/services/session/service/lifecycle/session.config.service.ts": 1080,
