+++ src/backend/services/session/service/chat/chat-message-handlers.service.ts
@@ -15,11 +15,10 @@ import type { SessionInitPolicyBridge } from '@/backend/services/session/service
 import { sessionDataService } from '@/backend/services/session/service/data/session-data.service';
 import { toErrorMessage } from '@/backend/services/session/service/lifecycle/session.error-message';
 import {
-  configureChatMessageHandlerLifecycle,
   sessionConfigService,
   sessionPermissionService,
   sessionService,
-} from '@/backend/services/session/service/lifecycle/session-services';
+} from '@/backend/services/session/service/lifecycle/session-core-services';
 import { sessionDomainService } from '@/backend/services/session/service/session-domain.service';
 import { VOICE_MODE_BREVITY_INSTRUCTION } from '@/backend/services/session/service/voice/voice-mode-instructions';
 import { workspaceNotificationService } from '@/backend/services/workspace';
@@ -904,6 +903,3 @@ export class ChatMessageHandlerService {
   // Private: Message handlers now live in chat-message-handlers/handlers
   // ============================================================================
 }
-
-export const chatMessageHandlerService = new ChatMessageHandlerService();
-configureChatMessageHandlerLifecycle(chatMessageHandlerService);
