+++ src/backend/services/session/service/chat/chat-message-handlers/handlers/set-config-option.handler.test.ts
@@ -6,7 +6,7 @@ const mocks = vi.hoisted(() => ({
   emitDelta: vi.fn(),
 }));
 
-vi.mock('@/backend/services/session/service/lifecycle/session-services', () => ({
+vi.mock('@/backend/services/session/service/lifecycle/session-core-services', () => ({
   sessionConfigService: {
     setSessionConfigOption: mocks.setSessionConfigOption,
     getSessionConfigOptions: mocks.getSessionConfigOptions,
