+++ src/backend/services/session/service/chat/chat-message-handlers/handlers/stop.handler.test.ts
@@ -7,7 +7,7 @@ const mocks = vi.hoisted(() => ({
   resetDispatchState: vi.fn(),
 }));
 
-vi.mock('@/backend/services/session/service/lifecycle/session-services', () => ({
+vi.mock('@/backend/services/session/service/lifecycle/session-core-services', () => ({
   sessionLifecycleService: {
     stopSession: mocks.stopSession,
   },
