+++ src/backend/services/session/service/index.ts
@@ -26,7 +26,7 @@ export type { SessionInitPolicyBridge, SessionWorkspaceBridge } from './bridges'
 // Chat services
 export type { EventForwarderContext } from './chat/chat-event-forwarder.service';
 export { chatEventForwarderService } from './chat/chat-event-forwarder.service';
-export { chatMessageHandlerService } from './chat/chat-message-handlers.service';
+export { ChatMessageHandlerService } from './chat/chat-message-handlers.service';
 export { sessionDataService } from './data/session-data.service';
 export { sessionProviderResolverService } from './data/session-provider-resolver.service';
 export type { SessionInterceptorBridge } from './interceptor.bridge';
@@ -47,6 +47,7 @@ export { SessionService } from './lifecycle/session.service';
 export type { SessionPromptService } from './lifecycle/session-services';
 export {
   acpEventProcessor,
+  chatMessageHandlerService,
   sessionConfigService,
   sessionLifecycleEventService,
   sessionLifecycleService,
