+++ src/backend/services/session/service/lifecycle/session-services.composition.test.ts
@@ -1,0 +1,45 @@
+import { afterEach, describe, expect, it, vi } from 'vitest';
+import { chatMessageHandlerService as publicChatMessageHandlerService } from '@/backend/services/session';
+import { WorkspaceStatus } from '@/shared/core';
+import { unsafeCoerce } from '@/test-utils/unsafe-coerce';
+import { chatMessageHandlerService, sessionLifecycleService } from './session-services';
+
+describe('session services composition', () => {
+  afterEach(() => {
+    vi.restoreAllMocks();
+  });
+
+  it('exports one configured chat singleton whose startup adapter filters chat-only options', async () => {
+    expect(publicChatMessageHandlerService).toBe(chatMessageHandlerService);
+    vi.spyOn(sessionLifecycleService, 'getSessionOptions').mockResolvedValue({
+      workingDir: '/tmp/worktree',
+      resumeProviderSessionId: undefined,
+      systemPrompt: undefined,
+      model: 'sonnet',
+      workspaceStatus: WorkspaceStatus.READY,
+    });
+    const getOrCreateSessionClient = vi
+      .spyOn(sessionLifecycleService, 'getOrCreateSessionClient')
+      .mockResolvedValue(unsafeCoerce({ sessionId: 'session-1' }));
+
+    await chatMessageHandlerService.handleMessage(
+      unsafeCoerce({ send: vi.fn() }),
+      'session-1',
+      '/tmp/worktree',
+      {
+        type: 'start',
+        thinkingEnabled: true,
+        planModeEnabled: true,
+        selectedModel: 'opus',
+        reasoningEffort: 'high',
+      }
+    );
+
+    expect(getOrCreateSessionClient).toHaveBeenCalledOnce();
+    expect(getOrCreateSessionClient).toHaveBeenCalledWith('session-1', {
+      thinkingEnabled: true,
+      model: 'opus',
+      reasoningEffort: 'high',
+    });
+  });
+});
